+++ mysql/tables/llx_commandedet.sql
@@ -32,10 +32,10 @@ create table llx_commandedet
   remise             real         DEFAULT 0,          -- montant de la remise
   fk_remise_except	 integer      NULL,               -- Lien vers table des remises fixes
   price              real,                            -- prix final
-  subprice           double(16,8) DEFAULT 0,          -- prix unitaire
-  total_ht           double(16,8) DEFAULT 0,          -- Total HT de la ligne toute quantité et incluant remise ligne et globale
-  total_tva          double(16,8) DEFAULT 0,          -- Total TVA de la ligne toute quantité et incluant remise ligne et globale
-  total_ttc          double(16,8) DEFAULT 0,          -- Total TTC de la ligne toute quantité et incluant remise ligne et globale
+  subprice           double(24,8) DEFAULT 0,          -- prix unitaire
+  total_ht           double(24,8) DEFAULT 0,          -- Total HT de la ligne toute quantité et incluant remise ligne et globale
+  total_tva          double(24,8) DEFAULT 0,          -- Total TVA de la ligne toute quantité et incluant remise ligne et globale
+  total_ttc          double(24,8) DEFAULT 0,          -- Total TTC de la ligne toute quantité et incluant remise ligne et globale
   info_bits		       integer      DEFAULT 0,          -- TVA NPR ou non
   marge_tx           double(6,3)  DEFAULT 0,          -- taux de marge (marge sur prix d'achat)
   marque_tx          double(6,3)  DEFAULT 0,          -- taux de marque (marge sur prix de vente)
